--- a/pilot/pkg/xds/delta_test.go
+++ b/pilot/pkg/xds/delta_test.go
@@ -398,9 +398,7 @@ func TestDeltaWDS(t *testing.T) {
 
 	// simulate a svc update
 	s.XdsUpdater.ConfigUpdate(&model.PushRequest{
-		ConfigsUpdated: sets.New(model.ConfigKey{
-			Kind: kind.Address, Name: svcA.ResourceName(), Namespace: svcA.Service.Namespace,
-		}),
+		AddressesUpdated: sets.New(svcA.ResourceName()),
 	})
 
 	resp = ads.ExpectResponse()
@@ -414,9 +412,7 @@ func TestDeltaWDS(t *testing.T) {
 	// simulate a svc delete
 	s.MemRegistry.RemoveServiceInfo(svcA)
 	s.XdsUpdater.ConfigUpdate(&model.PushRequest{
-		ConfigsUpdated: sets.New(model.ConfigKey{
-			Kind: kind.Address, Name: svcA.ResourceName(), Namespace: svcA.Service.Namespace,
-		}),
+		AddressesUpdated: sets.New(svcA.ResourceName()),
 	})
 
 	resp = ads.ExpectResponse()
@@ -429,18 +425,17 @@ func TestDeltaWDS(t *testing.T) {
 
 	// delete workload
 	s.MemRegistry.RemoveWorkloadInfo(wlA)
-	// a full push and a pod delete event
-	// This is a merged push request
+	// a pod delete event
 	s.XdsUpdater.ConfigUpdate(&model.PushRequest{
-		Full: true,
+		AddressesUpdated: sets.New(wlA.ResourceName()),
 	})
 
 	resp = ads.ExpectResponse()
 	if len(resp.RemovedResources) != 1 || resp.RemovedResources[0] != wlA.ResourceName() {
-		t.Fatalf("received unexpected removed eds resource %v", resp.RemovedResources)
+		t.Fatalf("received unexpected removed wds resource %v", resp.RemovedResources)
 	}
-	if len(resp.Resources) != 4 {
-		t.Fatalf("received unexpected eds resource %v", resp.Resources)
+	if len(resp.Resources) != 0 {
+		t.Fatalf("received unexpected wds resource %v", resp.Resources)
 	}
 }