Solves part 1 of https://github.com/istio/istio/issues/54499.

The idea here is that we use ConfigsUpdate which is lossy - at some points we wipe the configs, and the only choice is to push everything. At large scale, these full pushes cause dramatic spikes in Istio CPU utilization. For instance, imagine I have a massive scale of 100k pods and 1k ztunnels connected. At steady state, even with a crazy 10 pods/s changing, I need to push 10 workloads * 1k ztunnels = 10k workloads/s. If a full push is triggered, it will instantly need to push 100M workloads - a 10000x increase.

This fixes things by putting Addresses in a separate object. This _probably_ should not actually be specialized to Addresses and instead should also allow WADS resources maybe, to apply the same optimization there. However, in WADS we are using AuthorizationPolicy as the trigger kind so that may be awkward. Maybe we can make that its own type, too. Then instead of AddressesUpdated, this becomes more like `ConfigsThatHavePerfectDiffing` (we can pick a better name :-) ).

We still place objects into ConfigsUpdate. The reason is to avoid accidentally triggering a full push for the _other_ types; if we just sent a `Request{addressUpdated: [foo]}`, the other types will see `ConfigsUpdated=nil` and do a full push. We can likely improve this in the future, but for now this offers the best compatibility

Benchmark doesn't really capture this appropriately since we measure for fast a full/incremental push is; this fix makes full pushes less likely, so its not covered.
```
                                                           │  /tmp/master  │              /tmp/now              │
                                                           │    sec/op     │   sec/op     vs base               │
AddressFullGeneration/serviceentry-workloadentry-16           703.0µ ± 19%   698.3µ ± 7%        ~ (p=1.000 n=6)
AddressIncrementalGeneration/serviceentry-workloadentry-16   1035.0n ±  4%   856.4n ± 6%  -17.26% (p=0.002 n=6)
geomean                                                       26.97µ         24.45µ        -9.35%


                                                           │ /tmp/master  │              /tmp/now               │
                                                           │     B/op     │     B/op      vs base               │
AddressFullGeneration/serviceentry-workloadentry-16          869.0Ki ± 0%   849.1Ki ± 0%   -2.29% (p=0.002 n=6)
AddressIncrementalGeneration/serviceentry-workloadentry-16     968.0 ± 0%     776.0 ± 0%  -19.83% (p=0.002 n=6)
geomean                                                      28.66Ki        25.37Ki       -11.50%

                                                           │ /tmp/master │              /tmp/now              │
                                                           │  allocs/op  │  allocs/op   vs base               │
AddressFullGeneration/serviceentry-workloadentry-16          3.623k ± 0%   3.610k ± 0%   -0.36% (p=0.002 n=6)
AddressIncrementalGeneration/serviceentry-workloadentry-16    13.00 ± 0%    11.00 ± 0%  -15.38% (p=0.002 n=6)
geomean                                                       217.0         199.3        -8.18%
```

As part of this implementation, I have split out the on-demand vs wildcard code for WDS generator. I feel its much simpler to have 2 split codepaths vs having `ifs` all over the place. Actually, as part of this, I found some minor issues caused by the complexity of the `ifs` - not a meaningful bug, but `haveAliases` can never be used for instance.