@istio-testing 2025-01-08
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
lint_istio | 57c1d70bea748c31c98aba64d1d076ef58e90fa3 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/54601/lint_istio/1877034712078225408) | true | `/test lint`
integ-ipv6_istio | 57c1d70bea748c31c98aba64d1d076ef58e90fa3 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/54601/integ-ipv6_istio/1877034704121630720) | true | `/test integ-ipv6`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes-sigs/prow](https://github.com/kubernetes-sigs/prow/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->